Attempting to get rid of a container file rename operation in venus that tends to crash the win95 venus when the kernel still has the file referenced.

+++ coda-src/venus/venusrecov.cc
@@ -361,7 +361,7 @@ static void Recov_InitRVM() {
     if (InitMetaData)
     {
 	/* Compute recoverable storage requirements, and verify that log/data sizes are adequate. */
-        int RecovBytesNeeded = (int) RECOV_BYTES_NEEDED();
+        unsigned long RecovBytesNeeded = RECOV_BYTES_NEEDED();
 
         /* Set segment sizes if necessary. */
         if (VenusDataDeviceSize == UNSET_VDDS)
